requirements.txt
fastapi
pydantic
uvicorn
PyMuPDF
torch
numpy
sentence-transformers
faiss-cpu
transformers
accelerate
bitsandbytes